+++ packages/wds-engine/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.5](https://github.com/wanteddev/wds/compare/@wanteddev/wds-engine@1.0.4...@wanteddev/wds-engine@1.0.5) (2024-07-10)
+
+**Note:** Version bump only for package @wanteddev/wds-engine
+
 ## [1.0.4](https://github.com/wanteddev/wds/compare/@wanteddev/wds-engine@1.0.3...@wanteddev/wds-engine@1.0.4) (2024-05-21)
 
 **Note:** Version bump only for package @wanteddev/wds-engine